pyproject.toml
[project]
name = "serialportvisualization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bm3d>=4.0.3",
    "colormath>=3.0.0",
    "colour-science>=0.4.6",
    "deap>=1.4.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pwlf>=2.5.2",
    "pyyaml>=6.0.3",
    "scikit-image>=0.25.2",
    "scipy>=1.16.2",
]
